Search, Course, Run
Added final search functionality and fixed some bugs.
