+++ src/main/java/bugbusters/Course.java
@@ -146,7 +146,7 @@ public class Course {
         return id;
     }
 
-    private void setId(int id) {
+    public void setId(int id) {
         this.id = id;
     }
